Refactor : 페이지 컴포넌트 내부의 컴포넌트와 모듈 의존성 개선
- nav, buttons, feed 분류에 해당하는 컴포넌트 정리
- work.d.ts 파일에 work 라우트 내부에서 공유되는 타입 정리
- import 경로를 절대경로로 일괄변경 및 의존성 관계 정리
