+++ src/pages/work/work.d.ts
@@ -1,0 +1,36 @@
+export interface VideoCollection<T, U = T> {
+  filmShort: T;
+  outsource: U;
+}
+
+export interface WorkInfos {
+  title: string;
+  description: string;
+  resourceId: string;
+  category: string;
+  date: string;
+  thumbnailLink: string;
+  animatedThumbnailLink: string;
+}
+
+export interface VimeoVideos {
+  player_embed_url: string;
+  resource_key: string;
+  pictures: { base_link: string };
+  name: string;
+  description: string;
+  uri: string;
+  animated_thumbnail: string;
+}
+
+export interface OwnedVideoItems {
+  title: string;
+  category: VideosCategory;
+  date: string;
+  description: string;
+  resourceId: string;
+}
+
+export type ResourceHost = 'vimeo' | 'youtube';
+
+export type FlatformsCategory = 'filmShort' | 'outsource';
